gcc adjustments, still need some more investigations to get really clean
